refactor(release): reorganize sound utility imports and implement new validation and write functions
